% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_structure.R
\name{BlockDiagonal}
\alias{BlockDiagonal}
\title{Block diagonal matrix}
\usage{
BlockDiagonal(pk)
}
\arguments{
\item{pk}{vector encoding the grouping structure.}
}
\value{
A binary block diagonal matrix.
}
\description{
Generates a binary block diagonal matrix.
}
\examples{
# Example 1
BlockDiagonal(pk = c(2, 3))

# Example 2
BlockDiagonal(pk = c(2, 3, 2))
}
\seealso{
Other block matrix functions: 
\code{\link{BlockMatrix}()},
\code{\link{BlockStructure}()}
}
\concept{block matrix functions}
