% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{Contrast}
\alias{Contrast}
\title{Matrix contrast}
\usage{
Contrast(mat, digits = 3)
}
\arguments{
\item{mat}{input matrix.}

\item{digits}{number of digits to use.}
}
\value{
A single number, the contrast of the input matrix.
}
\description{
Computes matrix contrast, defined as the number of unique truncated
entries with a specified number of digits.
}
\examples{
# Example 1
mat <- matrix(c(0.1, 0.2, 0.2, 0.2), ncol = 2, byrow = TRUE)
Contrast(mat)

# Example 2
mat <- matrix(c(0.1, 0.2, 0.2, 0.3), ncol = 2, byrow = TRUE)
Contrast(mat)

}
\references{
\insertRef{ourstabilityselection}{fake}
}
