% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{SimulateSCM}
\alias{SimulateSCM}
\title{Data simulation for Structural Causal Modelling}
\usage{
SimulateSCM(
  n = 100,
  pk = c(5, 5, 5),
  theta = NULL,
  nu_between = 0.5,
  v_between = c(0.5, 1),
  v_sign = c(-1, 1),
  continuous = TRUE,
  residual_var = 1,
  scale = TRUE,
  output_matrices = FALSE
)
}
\arguments{
\item{n}{number of observations in the simulated data.}

\item{pk}{vector of the number of variables per group in the simulated data.
The number of nodes in the simulated graph is \code{sum(pk)}. With multiple
groups, the simulated (partial) correlation matrix has a block structure,
where blocks arise from the integration of the \code{length(pk)} groups.
This argument is only used if \code{theta} is not provided.}

\item{theta}{optional binary and symmetric adjacency matrix encoding the
conditional independence structure.}

\item{nu_between}{probability of having an edge between two nodes belonging
to different layers, as defined in \code{pk}. If \code{length(pk)=1}, this
is the expected density of the graph.}

\item{v_between}{vector defining the (range of) nonzero path coefficients. If
\code{continuous=FALSE}, \code{v_between} is the set of possible values. If
\code{continuous=TRUE}, \code{v_between} is the range of possible values.}

\item{v_sign}{vector of possible signs for path coefficients. Possible inputs
are: \code{1} for positive coefficients, \code{-1} for negative
coefficients, or \code{c(-1, 1)} for both positive and negative
coefficients.}

\item{continuous}{logical indicating whether to sample path coefficients from
a uniform distribution between the minimum and maximum values in
\code{v_between} (if \code{continuous=FALSE}) or from proposed values in
\code{v_between} (if \code{continuous=FALSE}).}

\item{residual_var}{vector of residual variances. The length of the vector
should be equal to \code{sum(pk)} or \code{ncol(theta)}. The same value is
used for all variables if a single value is provided (the default).}

\item{scale}{logical indicating if the true mean is zero and true variance is
one for all simulated variables. The observed mean and variance may be
slightly off by chance.}

\item{output_matrices}{logical indicating if the true path coefficients,
residual variances, and precision and (partial) correlation matrices should
be included in the output.}
}
\value{
A list with: \item{data}{simulated data with \code{n} observation and
  \code{sum(pk)} variables.} \item{theta}{adjacency matrix of the simulated
  Directed Acyclic Graph encoding causal relationships.}
  \item{path_coef}{simulated (true) matrix of path coefficients. Only
  returned if \code{output_matrices=TRUE}.} \item{omega}{simulated (true)
  precision matrix. Only returned if \code{output_matrices=TRUE}.}
  \item{phi}{simulated (true) partial correlation matrix. Only returned if
  \code{output_matrices=TRUE}.} \item{sigma}{simulated (true) covariance
  matrix. Only returned if \code{output_matrices=TRUE}.}
}
\description{
Simulates data from a Structural Causal Model (SCM). To ensure that the
generated SCM is identifiable, the nodes are organised by layers, with no
causal effects within layers.
}
\examples{
# Simulation of a layered SCM
set.seed(1)
pk <- c(3, 5, 4)
simul <- SimulateSCM(n = 100, pk = pk)
print(simul)
summary(simul)

# Visualisation of the layers
if (requireNamespace("igraph", quietly = TRUE)) {
  mygraph <- plot(simul)
  igraph::plot.igraph(mygraph,
    layout = igraph::layout_with_sugiyama(mygraph,
      layers = rep.int(1:length(pk), times = pk)
    )
  )
}
}
\seealso{
\code{\link{SimulatePrecision}}, \code{\link{MakePositiveDefinite}},
  \code{\link{Contrast}}

Other simulation functions: 
\code{\link{SimulateAdjacency}()},
\code{\link{SimulateClustering}()},
\code{\link{SimulateComponents}()},
\code{\link{SimulateGraphical}()},
\code{\link{SimulateRegression}()}
}
\concept{simulation functions}
