% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_structure.R
\name{BlockMatrix}
\alias{BlockMatrix}
\title{Block matrix}
\usage{
BlockMatrix(pk)
}
\arguments{
\item{pk}{vector encoding the grouping structure.}
}
\value{
A symmetric block matrix.
}
\description{
Generates a symmetric block matrix of size (\code{sum(pk)} x \code{sum(pk)}).
The sizes of the submatrices is defined based on \code{pk}. For each
submatrix, all entries are equal to the submatrix (block) index.
}
\examples{
# Example 1
BlockMatrix(pk = c(2, 3))

# Example 2
BlockMatrix(pk = c(2, 3, 2))
}
\seealso{
Other block matrix functions: 
\code{\link{BlockDiagonal}()},
\code{\link{BlockStructure}()}
}
\concept{block matrix functions}
